Lsp initates the truck - example not working with the deliver events
