+++ Server/Agents/Shipper.py
@@ -1,5 +1,5 @@
 from Agents.Misc import Request
-import Simulation
+from Simulation.Global import lsps
 
 class Shipper:
     def __init__(self, id):
@@ -14,9 +14,16 @@ class Shipper:
         quotes = []
         print("Shipper rules!")
         for lsp in self.lsp_list:
-            selected_lsp, quote = lsp.process_request(env, request)
-            quotes.append((selected_lsp, quote))
-        # selected_lsp, quote = min(quotes, key=lambda x: (yield x[1])[1])
+            selected_carrier, quote = lsp.process_request(env, request)
+            quotes.append((selected_carrier, quote))
         
+        selected_lsp = min(quotes, key=lambda x: x[1])[0]
+
+        lsps[selected_lsp].initiate_truck(env, request)
+
         # Schedule truck dispatch when time-window starts
         yield env.timeout(request.time_window[0] - env.now)
+
+    def delivered_request(self, env, request):
+        print(f"Request {request.id} delivered at time {env.now}")
+        self.requests.remove(request)
